security_NosymfollowMountOption: separate test for SELinux
This test needs to test the behavior of the chromiumos LSM code
alongside SELinux if and only if SELinux is enabled, so make a separate
test that does so. This separate test will throw a TestNAError if
SELinux is not installed on the board.

BUG=chromium:867807
TEST=ran autotest on board with/without SELinux enabled

Change-Id: I4c1a880b6ab7dbe3f0a6ff30ea401fc07aba7498
Reviewed-on: https://chromium-review.googlesource.com/1387765
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>