rlz: Increase time for the device to get locked.
When checking the rlz ping is sent in guest mode, we first need to login
as a regular user so that the device becomes locked for rlz.

We are seeing a lot of failures where the ping is not sent in 2 minutes.
We previously tried to increase that to 5 minutes but it had no effect.
So here I will give the test more time to get locked before switching to
guest mode.

BUG=None
TEST=None

Change-Id: I6d7ae4688838ecc35ea8acffc4e536b116387e5d
Reviewed-on: https://chromium-review.googlesource.com/1388184
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Stephen Lin <stephenlin@google.com>