[autotest] power_SuspendStress: make net-dev detection more forgiving
This code had several problems:
(1) it didn't handle the condition of "no devices" -- it would produce
an error in min() instead
(2) it doesn't do a good job of sorting our interfaces by metric
(min(my_dictionary) sorts by keys, not the values)
(3) it's not forgiving for thte case of a network device that comes back
soon after this test starts

For (1): short-circuit the detection when the dictionary is empty

For (2): add a 'key=' parameter to the use of min() (i.e., look at the
dictionary values, not the keys) and convert the 'Flags' field from a
string to int for better sorting.

For (3): add a polling loop. If we can't find a device after 10 seconds,
we're probably in bad shape and can bail out with an informative Error.

BUG=chromium:901876
TEST=power_SuspendStress, with and without good connectivity, and while
     forcing the eth0 link to go down

Change-Id: I91e6ab9dbf4162483c96adf953853752eb203da2
Reviewed-on: https://chromium-review.googlesource.com/c/1407804
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>