audio_Microphone: limit channel ids for Octopus
Octopus audio topologies list some extra recording channels which are
not present. We don't believe this has any user impact (and will be
fixed), but in the interim, this commit limits the max channel ID to 4
in this test so that the test will pass.

BUG=b:120568903
TEST=test_that audio_Microphone on an Octopus system

Change-Id: Iad886be06aafa1b85d26dfeb415c06d55566549c
Reviewed-on: https://chromium-review.googlesource.com/1418038
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>