platform_EncryptedStateful: Disable reclamation check for kernels < 4.4
With b/80549098, PUNCH_HOLE was disabled for all kernel trees before 4.4.
Disable the reclamation check for devices without the capability of
PUNCH_HOLE. For such devices, the only difference is that if a file is
deleted on the encrypted stateful partition, that space is not reclaimed
into the stateful partition and encrypted stateful mount will monotonically
grow upto the limit (30% of the stateful partition).

BUG=b:80549098
TEST=test_that platform_EncryptedStateful passed on Caroline(3.18).

Change-Id: I5ac4d97e9a70cf14c0dabb7c251a74e41c8c9ec6
Reviewed-on: https://chromium-review.googlesource.com/1432172
Commit-Ready: Sarthak Kukreti <sarthakkukreti@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>