faft: Add config option for lid wake capability
Add a config option for devices where the power architecture does
not allow for the lid to wake the system from a powered off state.

In this case it is necessary to press the power button after the
lid is opened.

The firmware_FwScreenCloseLid and power_BootUpTime tests are updated
to use this new flag.

BUG=b:121380403
TEST=pass firmware_FwScreenCloseLid and power_BootUpTime on Sarien

Change-Id: Ic7b12b14f18d3a544e96a6e6adc1d068e87840e7
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1432653