autotest: correctly initialize P test
When the tests were separated for N and P versions of ARC++, this test
had a call to super(...) that was not used in the other tests.
Subsequently running this test for P no longer worked since it was
referencing a class that no longer existed. Changing it to the new class
name resolves this problem.

BUG=b:117898487
TEST=test_that dut desktopui_ExitOnSupervisedUserCrash_P

Change-Id: Iad57b543412fb912f798fe8a285df2687f2deca4
Reviewed-on: https://chromium-review.googlesource.com/1437038
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Kenny Root <kroot@google.com>
Reviewed-by: Pohsien Wang <pwang@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>