TPMNotCorruptedDevMode: support factory mode DUTs
Add the kernel anti-rollback info state that has been observed on DUTs
in factory mode to the possible expected values.  Since FAFT is required
betweeen proto and EVT, we can not guarantee, from a process
perspective, that  DUTs that need to run this test are not in factory
mode).  This approach was selected additionally because doing something
like forcing a transition out of factory mode is non-trival and not
worth the effort considering the simplificiations that the TODO on the
expected values would introduce.

BUG=b:117409688
TEST=none
CQ-DEPEND=CL:1409823

Change-Id: I35862c84a526002b10d0f2ff6967abff6a2ba2d2
Reviewed-on: https://chromium-review.googlesource.com/1444731
Commit-Ready: Kevin Shelton <kmshelton@chromium.org>
Tested-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>