[Autotest] policy_ForceYouTubeRestrict
If the policy is set to strict than the user will not be able to view any
restricted videos on YouTube. If the policy is set to moderate than the
user will not be able to watch some restricted videos. In both cases
the user will not be able to toggle restricted settings on the website.

BUG=chromium:926306
TEST=this

Change-Id: I6a0d43df4a732be4b4dbd5a0b311896fa4b16320
Reviewed-on: https://chromium-review.googlesource.com/1443583
Commit-Ready: Ruben Zakarian <rzakarian@chromium.org>
Tested-by: Ruben Zakarian <rzakarian@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>