utils/run_pylint: Return the pylint error code instead of an exception
If a lint error was found, an exception would be thrown. This would
cause the program to end with a return code of 1, and it would print out
the cryptic error "ERROR:root:4" to stderr.

The pylint api specifies that the return code is a bit field:
* 0   no error
* 1   fatal message issued
* 2   error message issued
* 4   warning message issued
* 8   refactor message issued
* 16  convention message issued
* 32  usage error

This change makes it so the correct error code is returned, and the
exception is no longer thrown. The output on stderr was causing problems
when running pylint in sublime text because it expects stderr to be
empty.

BUG=none
TEST=Ran various tests

$ utils/run_pylint.py client/profilers/sar/sar.py
************* Module files.client.profilers.sar.sar
W:  7, 0: Unused import shutil (unused-import)
ERROR:root:4
$ echo $?
1

$ utils/run_pylint.py client/profilers/sar/sar.py
************* Module files.client.profilers.sar.sar
W:  7, 0: Unused import shutil (unused-import)
$ echo $?
4

$ utils/run_pylint.py client/profilers/
...
$ echo $?
20

Change-Id: I9284bd664095290c569fbe345ad95fd6078b47bc
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1446542
Reviewed-by: Jason Clinton <jclinton@chromium.org>