cr50_test: open cr50 from console if possible in fast_open
fast_open is not testing anything. It is just used to open cr50 as
quickly and reliably as possible. If OpenNoDevMode is accessible, skip
entering dev mode. If OpenFromUSB is accessible open from usb instead of
the AP. This reduces the number of things that are required to work to
get the device open and enable testlab mode.

BUG=none
BRANCH=none
TEST=run on soraka with prepvt and pvt images

Change-Id: I697785ca63d03da3abeabf9362019045f28a317f
Signed-off-by: Mary Ruthven <mruthven@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/1474493
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>