faft_cr50_prepvt: remove qual version
4.13 has issues. We are going to need to build 4.14. This change removes
the qual version for now so tests can keep running with 4.13 until we
build 4.14. Then we can update to 4.14 without waiting for a chromeos
image to build updating the qual version to 4.14.

The qual version was mostly added, so we can catch cleanup issues during
update test. The update tests are still part of experimental, so it's
not a very big risk right now. Waiting for chromeos builds sets us back
a lot more. We'll add this back later.

BUG=none
BRANCH=none
TEST=none

Change-Id: I4609c45b1167655ee0dc7ce1b36b9ccfc960872c
Reviewed-on: https://chromium-review.googlesource.com/c/1474494
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>