[autotest] site_sysinfo: don't collect .journal files
systemd-journald files are binary, and they aren't the preferred method
of log export. We have other logic that converts them to a text-based
format, which can stand up to the 'log_diff' parsing as well as the log
trimming we do on large files.

Skip these.

Unfortunately, the DEFAULT_EXCLUDES variable is written with rsync in
mind, and so we have to skip .journal files in two different ways.

BUG=chromium:923162, chromium:930226
TEST=test_that, see /var/log/journal/*.journal logs skipped

Change-Id: Ic44a0894cb0d0b7b5ff79894d4beb6f13702d7e0
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1493416
Reviewed-by: Aviv Keshet <akeshet@chromium.org>