Make telemetry_Crosperf work for new run_benchmark
The new run_benchmark provides a new profiling mechanism and this patch
will use it to replace the existing one, which provides more accurate
profile.

This patch is introduced after 1436196 patch checked in.

Note that this patch will only work for situation that only one perf
data is generated from the benchmark. Multiple story-tags is not
supported in crosperf yet.

TEST=tested with octane and webrtc with story-tag on eve
BUG=chromium:923145

Change-Id: I1d3938f8dfd072795277d2db7fad1de7a23989cd
Reviewed-on: https://chromium-review.googlesource.com/1419241
Commit-Ready: Zhizhou Yang <zhizhouy@google.com>
Tested-by: Zhizhou Yang <zhizhouy@google.com>
Reviewed-by: Caroline Tice <cmtice@chromium.org>