tast: Increase result size to 50MB.
Currently, it is throttled up to 20MB, and as its result,
log files are trimmed unexpectedly.

This increases the size to 50MB for crosbolt, informational and mustpass
Tast tests.
Because Tast consolidates multiple tests as if it is one autotest task,
the result size is the accumulation of them.

Note that, example and skylab-stage runs just a few tests,
so they are out of focus in this CL.

BUG=chromium:934090
TEST=Ran try.

Change-Id: I7d81cafceca9cb37d3bb7e2fe29edef83072813a
Reviewed-on: https://chromium-review.googlesource.com/1483550
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>