cheets_{CTS,GTS}: Configure retry count based on branch number.
Previous implementation that parsed CHROMEOS_RELEASE_DESCRIPTION did not
return the right channel, and hence we had to add hack for release branches
to adjust the retry count.

Instead, this new implementation uses the "branch number" field of
CHROMEOS_RELEASE_VERSION to determine the count. Besides, it splits
the configuration for the default retry count and maximum retry count.
(We have control files with max_retry=12 or max_retry=30 with the
intention to do that number of retries, but it was always capped to 9.)

BUG=b:119521922
TEST=cheets_CTS_N, cheets_CTS_P, cheets_GTS, cheets_CTS_Instant on
  dev/beta/stable devices all runs without problems
CQ-DEPEND=CL:*905295

Change-Id: Ic76baf55456988ba988a78fb7f8ae57848f6f6d4
Reviewed-on: https://chromium-review.googlesource.com/1491092
Commit-Ready: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>