network_WiFi_IBSS: "N/A" without client support
We don't have good testing for ad-hoc networks and are likely dropping
it entirely. At least don't fail the test -- just mark it as N/A.

Dirty secret: we mostly are "N/A" today already, because many of the
test APs (whirlwind and gale) don't support this either, so we already
raise TestNA at the router capability check.

BUG=chromium:931442
TEST=run it, with and without ad-hoc support on the client

Change-Id: I5f48960b9bc3383316573b70240693ff5235a88f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1509917
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>