[autotest] Fix policy check in Enterprise Base
Changes to the chrome://policy page cause the policy check to error.

This CL is an addition to the already merged fix. This specifically
fixes the issue where the policy overflows the value column and requires
the Value row below. Additionally this will now allow extension tables
to be read.

BUG=chromium:937315
TEST=AllowDinosaurEasterEgg, EnterpriseForceInstall, ExtensionPolicy

Change-Id: If69fcc4d486a5e1985a021376f689d433ef19e56
Reviewed-on: https://chromium-review.googlesource.com/1516232
Commit-Ready: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>