firmware_EventLog: Allow S0ix event log entries in suspend/resume test
Instead of forcing S3 for this test, allow S0ix events to be valid and
use the default/configured method for suspend/resume.  If the test fails
with the default(s0ix) method then disable suspend_to_idle and attempt
to force S3 mode.

This should have no impact on currently passing boards because if the
check for s0ix event logs fails and will try again with S3.

BUG=b:128789923
TEST=pass firwmare_EventLog on Sarien once s0ix logging is enabled

Change-Id: I74a2a91d64117cf623d9e9d0824d75b804d1f878
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1526873
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>