deploy: force create_host as a contextmanager
CrosHost has a `close` method which destroys master ssh process, dumps
UART logs etc. But unfortunately, we don't call it all the time. This
caused some problems, e.g. crbug.com/803682. However we shouldn't be
blamed because it's a bad idea to ask us to remember call it explicitly
every time after creating the host object.

There are two ways to resolve this problem:
1. Call `close` in __del__ method, so it can be called automatically.
However, we cannot guarantee the execution order, and `close` method of
embeded objects may be called multiple times.

2. Enforce a context manager to create the host object. This way is
better as we can control when to close the object and it's more
pythonic.

I prefer #2. And I think the deploy script is a good start to use it.
That's what this CL does.

BUG=chromium:943673
TEST=Run both deploy script and prepare/main.py locally.

Change-Id: I35cd771d6c9872f0e63dfaedc0bda08cfe90ec5e
Reviewed-on: https://chromium-review.googlesource.com/1534253
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Congbin Guo <guocb@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>