P2P tests: Retry one, make another less strict.
Sometimes when the 2nd DUT looks for the 1st DUT to update via P2P it
will find another DUT in the lab with P2P enabled and use that instead
of the DUT we expect. I have been failing the test in this case but I
now I think it is OK since it is still actually updating via p2p and we
cannot control which DUT P2P decides to use.

I've added a retry to the P2P test for another issue where it cannot
find any hosts sometimes.

BUG=None
TEST=tested it locally

Change-Id: I179dccd9b6009d24d8f0224c57cc9619bbbd012a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1534252
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>