[autotest][cfm] Aggregate result of StorageWrittenMetric
The deviation for each sample is too high and is only causing extra
noise in the graphs. Also report the total sum as a single sample.

To enable this, Metric has been extended with an aggregate method that
subclasses can implement. This will be called when one tries to get the
values of the metric.

BUG=chromium:933363
TEST=Unit tests and local device

Change-Id: I118663826b3cfa1fa1c8a68e99323620e46599e5
Reviewed-on: https://chromium-review.googlesource.com/1528980
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Emil Lundmark <lndmrk@chromium.org>
Reviewed-by: Denis Tosic <dtosic@google.com>