g2f_utils: Add p1 parameter to register/auth functions
A recent change to g2ftool makes it by default pass
a value of P1 that requires user presence, meaning that
tests must simulate this with servo.

This allows the P1 value to be specified, with a default
value of 0, which was the previous default value in
g2ftool.

Also, removes some unused imports.

BUG=chromium:945932
TEST=test_that .. firmware_Cr50U2fPowerwash

Change-Id: I61698bb9b7edd15071608cbdfeab9d9aca0a6d6d
Reviewed-on: https://chromium-review.googlesource.com/1539445
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Louis Collard <louiscollard@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>