run_suite_skylab: error out if suite's swarming task id not known
Suites with no defined suite_id shouldn't be created, because they
result in tasks that can't be waited on via the usual wait mechanism.
This is perplexing when attempting to test run_suite_skylab locally.

Making the failure explicit makes it easier to understand and test.

BUG=None
TEST=None

Change-Id: I3718efb9b99949fa361be95cf29fc268cd019520
Reviewed-on: https://chromium-review.googlesource.com/1544890
Commit-Ready: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>